src/google-classroom-extractor/pyproject.toml
[tool.poetry]
name = "efi-google-classroom-extractor"
version = "1.0.0-alpha.0"
homepage = "https://techdocs.ed-fi.org/display/EDFITOOLS/LMS+Toolkit"
repository = "https://github.com/Ed-Fi-Exchange-OSS/Ed-Fi-X-Fizz"
description = "Extract tool for retrieving student data from Google Classroom"
authors = ["Ed-Fi Alliance, LLC, and contributors"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^1.11.0"
google-auth-oauthlib = "^0.4.1"
pandas = "^1.1.1"
python-dotenv = "^0.14.0"
opnieuw = "^1.1.0"
tail-recursive = "^2.0.0"
SQLAlchemy = "^1.3.19"
teamcity-messages = "^1.28"
pytest = "^6.0"
xxhash = "^2.0.0"
numpy = "1.19.3"
errorhandler = "^2.0.1"
ConfigArgParse = "^1.2.3"
edfi-lms-extractor-lib = "1.0.0a0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
pylint = "^2.6.0"
black = "^20.8b1"
flake8 = "^3.8.3"
faker = "^4.1.3"
coverage = "^5.3"
mypy = "^0.782"
pytest-describe = "^1.0.0"
pook = "^1.0.1"
keyring = "^22.0.1"
twine = "^3.3.0"

[build-system]
requires = ["poetry>=1.0.10"]
build-backend = "poetry.masonry.api"
